requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0
gunicorn==21.2.0
python-dotenv==1.0.0